requirements.txt
certifi==2020.12.5
chardet==4.0.0
configparser==5.0.1
idna==2.10
python-dotenv==0.21.0
requests==2.25.1
six==1.15.0
spotipy==2.16.1
urllib3==1.26.2
